backend/requirements.txt
# Django
Django>=5.0,<6.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.3,<5.0

# Authentication
django-allauth>=0.61,<1.0
dj-rest-auth>=5.0,<6.0

# Database
psycopg2-binary>=2.9,<3.0

# Images
Pillow>=11.0

# Rich text editor
django-ckeditor-5>=0.2,<1.0

# Filtering
django-filter>=23.5,<24.0

# Static files
whitenoise>=6.6,<7.0

# Environment variables
python-dotenv>=1.0,<2.0

# Development
django-debug-toolbar>=4.2,<5.0
